move environament check into env module
